refactor whole package 2
